requirements.txt
requests==2.33.1
websocket-client==1.9.0
PyExecJS==1.5.1
loguru==0.7.2
protobuf==4.25.3
flask==2.3.3
flask-cors==4.0.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pytest==7.4.3
pytest-flask==1.2.0
python-dateutil==2.8.2
playwright==1.42.0